+++ CHANGELOG.md
@@ -2,27 +2,46 @@
 
 ## Version History
 
-### v3.4.0
+### v3.5.2
+- Bug fixes
+
 ---
+
+### v3.5.1
+- ADDED user content moderation callback
+
+---
+
+### v3.5.0
+- ADDED support for deeplinking
+- ADDED invite data bundling
+
+---
+
+### v3.4.0
 - Support for activities with image, button, action
 - onGameAvatarClick handler
 - onActivityActionClick handler
 - Leaderboard data exposed to developer
 - Chat Kill switch
 
-### v3.3.0
 ---
+
+### v3.3.0
 - Global Chat Room
 
-### v3.2.0
 ---
+
+### v3.2.0
 - Updated chat engine
 - Small improvements
 
-### v3.1.1
 ---
+
+### v3.1.1
 - Small bug fixes
 
-### v3.1.0 
 ---
+
+### v3.1.0 
 - Initial GetSocial version